Moved out utils, docs and loggers to commons module
