+++ micrometer-commons/src/main/java/io/micrometer/common/ImmutableTag.java
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micrometer.observation;
+package io.micrometer.common;
 
 import java.util.Objects;
 
-import io.micrometer.observation.lang.Nullable;
+import io.micrometer.common.lang.Nullable;
 
 import static java.util.Objects.requireNonNull;
