+++ micrometer-commons/src/main/java/io/micrometer/common/Tag.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micrometer.observation;
+package io.micrometer.common;
 
 /**
  * Key/value pair representing a dimension of a meter used to classify and drill into measurements.
