+++ micrometer-commons/src/main/java/io/micrometer/common/lang/Nullable.java
@@ -1,0 +1,52 @@
+/*
+ * Copyright 2002-2017 the original author or authors.
+ * Copyright 2017-2021 VMware, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.micrometer.common.lang;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import javax.annotation.Nonnull;
+import javax.annotation.meta.TypeQualifierNickname;
+import javax.annotation.meta.When;
+
+/**
+ * A common annotation to declare that annotated elements can be {@code null} under
+ * some circumstance. Leverages JSR 305 meta-annotations to indicate nullability in Java
+ * to common tools with JSR 305 support and used by Kotlin to infer nullability of the API.
+ * <p>Should be used at parameter, return value, and field level. Methods override should
+ * repeat parent {@code @Nullable} annotations unless they behave differently.
+ * <p>Can be used in association with {@code NonNullApi} or {@code @NonNullFields} to
+ * override the default non-nullable semantic to nullable.
+ * <p>
+ * NOTE: This file has been copied from {@code org.springframework.lang}.
+ *
+ * @author Sebastien Deleuze
+ * @author Juergen Hoeller
+ * @see NonNullApi
+ * @see NonNullFields
+ * @see NonNull
+ */
+@Target({ElementType.METHOD, ElementType.PARAMETER, ElementType.FIELD})
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+@Nonnull(when = When.MAYBE)
+@TypeQualifierNickname
+public @interface Nullable {
+}
