+++ micrometer-commons/src/main/java/io/micrometer/common/util/internal/logging/JdkLogger.java
@@ -52,7 +52,7 @@
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  *
  */
-package io.micrometer.observation.util.internal.logging;
+package io.micrometer.common.util.internal.logging;
 
 import java.util.logging.Level;
 import java.util.logging.LogRecord;
