+++ micrometer-commons/src/main/java/io/micrometer/common/util/internal/logging/WarnThenDebugLogger.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micrometer.observation.util.internal.logging;
+package io.micrometer.common.util.internal.logging;
 
 import java.util.concurrent.atomic.AtomicBoolean;
