+++ micrometer-observation/build.gradle
@@ -1,6 +1,7 @@
 description 'Module containing Observation related code'
 
 dependencies {
+    api project(":micrometer-commons")
 
     optionalApi 'io.micrometer:context-propagation-api'
 
@@ -8,8 +9,8 @@ dependencies {
     optionalApi 'javax.servlet:javax.servlet-api'
 
     // log monitoring
-    optionalApi 'ch.qos.logback:logback-classic'
-    optionalApi 'org.apache.logging.log4j:log4j-core'
+    testImplementation 'ch.qos.logback:logback-classic'
+    testImplementation 'org.apache.logging.log4j:log4j-core'
 
     testImplementation project(':micrometer-observation-test')
