+++ micrometer-observation/src/main/java/io/micrometer/observation/NoopObservation.java
@@ -16,6 +16,8 @@
 package io.micrometer.observation;
 
 
+import io.micrometer.common.Tag;
+
 /**
  * No-op implementation of {@link Observation} so that we can disable the instrumentation logic.
  *
