+++ micrometer-observation/src/main/java/io/micrometer/observation/ObservationTextPublisher.java
@@ -19,7 +19,7 @@ import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.function.Predicate;
 
-import io.micrometer.observation.util.internal.logging.InternalLoggerFactory;
+import io.micrometer.common.util.internal.logging.InternalLoggerFactory;
 
 /**
  * An {@link ObservationHandler} that converts the context to text and Publishes it to the {@link Consumer} of your choice.
