+++ micrometer-observation/src/main/java/io/micrometer/observation/transport/http/tags/DefaultHttpTagsProvider.java
@@ -15,7 +15,7 @@
  */
 package io.micrometer.observation.transport.http.tags;
 
-import io.micrometer.observation.Tags;
+import io.micrometer.common.Tags;
 import io.micrometer.observation.transport.http.HttpRequest;
 import io.micrometer.observation.transport.http.HttpResponse;
